+++ lib/l10n/app_localizations.dart
@@ -4294,6 +4294,12 @@ abstract class AppLocalizations {
   /// **'Tags'**
   String get tags;
 
+  /// No description provided for @addTag.
+  ///
+  /// In en, this message translates to:
+  /// **'Add a tag'**
+  String get addTag;
+
   /// No description provided for @subscription.
   ///
   /// In en, this message translates to:
@@ -5002,12 +5008,6 @@ abstract class AppLocalizations {
   /// **'Main amount in the selected currency.'**
   String get tooltipFieldAmount;
 
-  /// No description provided for @tooltipFieldCurrency.
-  ///
-  /// In en, this message translates to:
-  /// **'Primary currency of this split.'**
-  String get tooltipFieldCurrency;
-
   /// No description provided for @tooltipFieldForeignAmount.
   ///
   /// In en, this message translates to:
@@ -5041,7 +5041,7 @@ abstract class AppLocalizations {
   /// No description provided for @tooltipFieldTags.
   ///
   /// In en, this message translates to:
-  /// **'Comma-separated tags for quick filtering.'**
+  /// **'Tags for quick filtering. Enter adds one, the cross takes it off.'**
   String get tooltipFieldTags;
 
   /// No description provided for @tooltipFieldSubscription.